Modularize meeting accessibility tests
Separate participant, Slack, inspection, chat, capture, parsing, and browser coverage into focused macOS suites while preserving accessibility behavior.
