+++ crates/detect/src/meeting_ax.rs
@@ -2555,1838 +2555,4 @@ impl From<cg::Rect> for AxRect {
 }
 
 #[cfg(all(test, target_os = "macos"))]
-mod tests {
-    use super::*;
-
-    fn node(index: usize, role: &str, title: &str, bounds: Option<AxRect>) -> AxNode {
-        AxNode {
-            index,
-            tree_path: vec![index],
-            element_hash: None,
-            role: Some(role.to_string()),
-            identifier: None,
-            title: Some(title.to_string()),
-            value: None,
-            description: None,
-            placeholder: None,
-            enabled: Some(true),
-            settable_value: false,
-            bounds,
-            text: node_text(
-                &Some(role.to_string()),
-                &Some(title.to_string()),
-                &None,
-                &None,
-                &None,
-            ),
-            within_zoom_meeting_scope: false,
-            within_zoom_chat_scope: false,
-            within_slack_huddle_scope: false,
-        }
-    }
-
-    fn fixture_node(index: usize, role: &str, title: &str, path: &[usize]) -> AxNode {
-        let mut node = node(
-            index,
-            role,
-            title,
-            Some(AxRect {
-                x: 10.0,
-                y: 10.0,
-                width: 120.0,
-                height: 40.0,
-            }),
-        );
-        node.tree_path = path.to_vec();
-        node.element_hash = Some(0x1000 + index);
-        node
-    }
-
-    fn fixture_composer(index: usize, title: &str, path: &[usize]) -> AxNode {
-        let mut node = fixture_node(index, "AXTextArea", title, path);
-        node.settable_value = true;
-        node
-    }
-
-    fn ancestor(label: &str) -> AxAncestor {
-        ancestor_at(label, &[0])
-    }
-
-    fn ancestor_at(label: &str, path: &[usize]) -> AxAncestor {
-        AxAncestor {
-            path: path.to_vec(),
-            labels: vec![label.to_string()],
-        }
-    }
-
-    fn zoom_message_node(index: usize, text: &str) -> AxNode {
-        let mut node = node(index, "AXStaticText", text, None);
-        node.within_zoom_meeting_scope = true;
-        node.within_zoom_chat_scope = true;
-        node
-    }
-
-    #[test]
-    fn test_participant_name_from_zoom_video_render_label() {
-        assert_eq!(
-            participant_name_from_evidence(
-                &MeetingPlatform::Zoom,
-                "Video render Ada Lovelace, Computer audio unmuted",
-            ),
-            Some("Ada Lovelace".to_string())
-        );
-    }
-
-    #[test]
-    fn test_zoom_video_render_becomes_stream_candidate_with_audio_state() {
-        let nodes = vec![node(
-            7,
-            "AXGroup",
-            "Video render Ada Lovelace, Computer audio unmuted",
-            Some(AxRect {
-                x: 0.0,
-                y: 0.0,
-                width: 320.0,
-                height: 180.0,
-            }),
-        )];
-
-        let streams =
-            find_participant_streams(&MeetingPlatform::Zoom, &MeetingSurface::Native, &nodes);
-
-        assert_eq!(streams.len(), 1);
-        assert_eq!(
-            streams[0].participant_name,
-            Some("Ada Lovelace".to_string())
-        );
-        assert!(!streams[0].is_active_speaker);
-        assert!(streams[0].confidence > 0.6);
-        assert!(
-            streams[0]
-                .signals
-                .contains(&"audio-state-label".to_string())
-        );
-    }
-
-    #[test]
-    fn test_explicit_active_speaker_label_marks_stream_active() {
-        let nodes = vec![node(
-            9,
-            "AXGroup",
-            "Video render Grace Hopper, active speaker",
-            Some(AxRect {
-                x: 0.0,
-                y: 0.0,
-                width: 320.0,
-                height: 180.0,
-            }),
-        )];
-
-        let streams =
-            find_participant_streams(&MeetingPlatform::Zoom, &MeetingSurface::Native, &nodes);
-
-        assert_eq!(streams.len(), 1);
-        assert!(streams[0].is_active_speaker);
-        assert!(
-            streams[0]
-                .signals
-                .contains(&"speaker-state-label".to_string())
-        );
-    }
-
-    #[test]
-    fn test_self_view_speaker_label_marks_stream_active() {
-        let nodes = vec![node(10, "AXRow", "Grace Hopper is speaking (You)", None)];
-
-        let streams =
-            find_participant_streams(&MeetingPlatform::Zoom, &MeetingSurface::Native, &nodes);
-
-        assert_eq!(streams.len(), 1);
-        assert_eq!(streams[0].participant_name.as_deref(), Some("Grace Hopper"));
-        assert!(streams[0].is_active_speaker);
-    }
-
-    #[test]
-    fn test_active_speaker_is_retained_past_participant_limit() {
-        let mut nodes = (0..24)
-            .map(|index| {
-                node(
-                    index,
-                    "AXGroup",
-                    &format!("Video render Participant {index}, Computer audio unmuted"),
-                    Some(AxRect {
-                        x: 0.0,
-                        y: 0.0,
-                        width: 320.0,
-                        height: 180.0,
-                    }),
-                )
-            })
-            .collect::<Vec<_>>();
-        nodes.push(node(24, "AXRow", "Ada Lovelace is speaking", None));
-
-        let streams =
-            find_participant_streams(&MeetingPlatform::Zoom, &MeetingSurface::Native, &nodes);
-
-        assert_eq!(streams.len(), 24);
-        assert!(streams.iter().any(|stream| {
-            stream.is_active_speaker && stream.participant_name.as_deref() == Some("Ada Lovelace")
-        }));
-    }
-
-    #[test]
-    fn test_zoom_prefers_named_speaker_state_over_generic_video_tile() {
-        let mut tile = fixture_node(9, "AXGroup", "Video tile", &[0, 1]);
-        tile.description = Some("Grace Hopper is speaking".to_string());
-        tile.text = node_text(
-            &tile.role,
-            &tile.title,
-            &tile.value,
-            &tile.description,
-            &tile.placeholder,
-        );
-
-        let streams =
-            find_participant_streams(&MeetingPlatform::Zoom, &MeetingSurface::Native, &[tile]);
-
-        assert_eq!(streams.len(), 1);
-        assert_eq!(
-            streams[0].participant_name,
-            Some("Grace Hopper".to_string())
-        );
-        assert_eq!(streams[0].id, "ax-element-1009");
-        assert!(streams[0].is_active_speaker);
-    }
-
-    #[test]
-    fn test_zoom_speaker_flag_uses_the_same_label_as_participant_name() {
-        let mut roster = fixture_node(
-            15,
-            "AXStaticText",
-            "Ada Lovelace (Host, me, Participant ID:417329) No audio connected",
-            &[0, 1],
-        );
-        roster.description = Some("Grace Hopper is speaking".to_string());
-        roster.text = node_text(
-            &roster.role,
-            &roster.title,
-            &roster.value,
-            &roster.description,
-            &roster.placeholder,
-        );
-
-        let stream = candidate_stream(&MeetingPlatform::Zoom, &MeetingSurface::Native, &roster)
-            .expect("expected Zoom roster participant");
-
-        assert_eq!(stream.participant_name.as_deref(), Some("Ada Lovelace"));
-        assert!(!stream.is_active_speaker);
-        assert!(!stream.signals.contains(&"speaker-state-label".to_string()));
-    }
-
-    #[test]
-    fn test_participant_streams_deduplicate_repeated_named_ax_nodes() {
-        let first = fixture_node(9, "AXGroup", "Grace Hopper is speaking", &[0, 1]);
-        let second = fixture_node(10, "AXRow", "Grace Hopper is speaking", &[0, 1, 0]);
-
-        let streams = find_participant_streams(
-            &MeetingPlatform::Zoom,
-            &MeetingSurface::Native,
-            &[first, second],
-        );
-
-        assert_eq!(streams.len(), 1);
-        assert_eq!(
-            streams[0].participant_name,
-            Some("Grace Hopper".to_string())
-        );
-    }
-
-    #[test]
-    fn test_participant_streams_keep_distinct_people_with_the_same_name() {
-        let first = fixture_node(11, "AXGroup", "Alex Kim is speaking", &[0, 1]);
-        let second = fixture_node(12, "AXGroup", "Alex Kim is speaking", &[0, 2]);
-
-        let streams = find_participant_streams(
-            &MeetingPlatform::Zoom,
-            &MeetingSurface::Native,
-            &[first, second],
-        );
-
-        assert_eq!(streams.len(), 2);
-        assert!(
-            streams
-                .iter()
-                .all(|stream| stream.participant_name.as_deref() == Some("Alex Kim"))
-        );
-    }
-
-    #[test]
-    fn test_slack_profile_row_is_participant_without_claiming_audio_is_speaking() {
-        let streams = find_participant_streams(
-            &MeetingPlatform::Slack,
-            &MeetingSurface::Native,
-            &[
-                node(12, "AXCell", "View John Jeong's profile", None),
-                node(13, "AXStaticText", "Audio", None),
-            ],
-        );
-
-        assert_eq!(streams.len(), 1);
-        assert_eq!(streams[0].participant_name, Some("John Jeong".to_string()));
-        assert!(!streams[0].is_active_speaker);
-    }
-
-    #[test]
-    fn test_slack_huddle_requires_huddle_label_and_enabled_leave_control() {
-        let mut composer = node(2, "AXTextArea", "Message to test", None);
-        composer.settable_value = true;
-        let mut disabled_leave = node(1, "AXButton", "Leave Huddle", None);
-        disabled_leave.enabled = Some(false);
-
-        assert_eq!(slack_huddle_context(&[composer.clone()]), None);
-        assert_eq!(
-            slack_huddle_context(&[node(0, "AXWindow", "Huddle in test", None), disabled_leave,]),
-            None
-        );
-        assert_eq!(
-            slack_huddle_context(&[
-                node(0, "AXWindow", "Huddle in  test", None),
-                node(1, "AXButton", "Leave Huddle", None),
-                composer,
-            ]),
-            Some(("Huddle in  test".to_string(), "test".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_slack_live_huddle_controls_fit_tree_depth_budget() {
-        assert!(MAX_TREE_DEPTH >= 14);
-    }
-
-    #[test]
-    fn test_ordinary_slack_composer_is_not_a_huddle_composer() {
-        let mut composer = node(2, "AXTextArea", "Message #general", None);
-        composer.settable_value = true;
-
-        assert!(!is_slack_huddle_composer(&composer, "test"));
-        assert!(candidate_chat_target(&composer).is_none());
-        assert_eq!(slack_huddle_context(&[composer]), None);
-    }
-
-    #[test]
-    fn test_slack_hidden_thread_control_is_recognized() {
-        let control = node(3, "AXButton", "Show/hide Thread", None);
-
-        assert!(is_slack_thread_control(&control));
-        assert_eq!(
-            candidate_chat_target(&control).unwrap().kind,
-            "openChatControl"
-        );
-    }
-
-    #[test]
-    fn test_slack_composer_and_send_button_must_share_live_thread_container() {
-        let mut composer = node(4, "AXTextArea", "Message to test", None);
-        composer.settable_value = true;
-        let thread = [
-            ancestor("Thread in test (private channel)"),
-            ancestor("composer"),
-        ];
-        let other_thread = [ancestor("Thread in random (private channel)")];
-        let duplicate_label_other_path = [ancestor_at("Thread in test (private channel)", &[9, 4])];
-
-        assert!(is_slack_huddle_composer_in_thread(
-            &composer, &thread, "test"
-        ));
-        assert!(!is_slack_huddle_composer_in_thread(&composer, &[], "test"));
-
-        let mut send = node(5, "AXButton", "Send now", None);
-        send.enabled = Some(false);
-        assert!(!is_slack_send_now_in_thread(&send, &thread, "test", &[0]));
-
-        send.enabled = Some(true);
-        assert!(is_slack_send_now_in_thread(&send, &thread, "test", &[0]));
-        assert!(!is_slack_send_now_in_thread(
-            &send,
-            &other_thread,
-            "test",
-            &[0]
-        ));
-        assert!(!is_slack_send_now_in_thread(
-            &send,
-            &duplicate_label_other_path,
-            "test",
-            &[0]
-        ));
-    }
-
-    #[test]
-    fn test_slack_composer_selection_fails_on_ambiguity_and_drafts() {
-        let mut first = node(4, "AXTextArea", "Message to test", None);
-        first.settable_value = true;
-        let mut second = first.clone();
-        second.index = 5;
-
-        assert_eq!(
-            unique_matching_index([&first, &second].into_iter().enumerate(), |node| {
-                is_slack_huddle_composer(node, "test")
-            },),
-            UniqueMatch::Ambiguous
-        );
-
-        first.value = Some("existing draft".to_string());
-        assert!(has_nonempty_draft(&first));
-        first.value = Some(" \n ".to_string());
-        assert!(!has_nonempty_draft(&first));
-        assert!(chat_input_is_owned("disclosure", "disclosure"));
-        assert!(!chat_input_is_owned(
-            "disclosure plus user text",
-            "disclosure"
-        ));
-    }
-
-    #[test]
-    fn test_chat_inspection_does_not_use_writable_value_as_label() {
-        let mut input = node(6, "AXTextArea", "", None);
-        input.title = None;
-        input.settable_value = true;
-        input.value = Some("private draft".to_string());
-        input.text = searchable_node_text(
-            &input.role,
-            &input.title,
-            &input.value,
-            &input.description,
-            &input.placeholder,
-            input.settable_value,
-        );
-
-        assert!(!input.text.contains("private draft"));
-        assert!(candidate_chat_target(&input).is_none());
-        assert_eq!(inspection_label(&input), None);
-
-        let mut read_only_input = input.clone();
-        read_only_input.settable_value = false;
-        read_only_input.value = Some("private read-only text".to_string());
-        read_only_input.text = searchable_node_text(
-            &read_only_input.role,
-            &read_only_input.title,
-            &read_only_input.value,
-            &read_only_input.description,
-            &read_only_input.placeholder,
-            read_only_input.settable_value,
-        );
-        assert!(!read_only_input.text.contains("private read-only text"));
-        assert_eq!(node_labels(&read_only_input).count(), 0);
-
-        let mut participant = node(
-            7,
-            "AXImage",
-            "",
-            Some(AxRect {
-                x: 0.0,
-                y: 0.0,
-                width: 200.0,
-                height: 100.0,
-            }),
-        );
-        participant.title = None;
-        participant.settable_value = true;
-        participant.value = Some("private participant value".to_string());
-        participant.text = searchable_node_text(
-            &participant.role,
-            &participant.title,
-            &participant.value,
-            &participant.description,
-            &participant.placeholder,
-            participant.settable_value,
-        );
-        assert!(
-            candidate_stream(
-                &MeetingPlatform::Slack,
-                &MeetingSurface::Native,
-                &participant,
-            )
-            .is_none()
-        );
-
-        let mut secure_input = read_only_input.clone();
-        secure_input.role = Some("AXSecureTextField".to_string());
-        secure_input.value = Some("private password".to_string());
-        secure_input.text = searchable_node_text(
-            &secure_input.role,
-            &secure_input.title,
-            &secure_input.value,
-            &secure_input.description,
-            &secure_input.placeholder,
-            secure_input.settable_value,
-        );
-        assert!(!secure_input.text.contains("private password"));
-        assert_eq!(node_labels(&secure_input).count(), 0);
-    }
-
-    #[test]
-    fn test_private_video_text_and_large_images_are_not_participant_streams() {
-        let private_nodes = [
-            node(20, "AXStaticText", "Private video notes", None),
-            node(
-                21,
-                "AXImage",
-                "Private camera preview",
-                Some(AxRect {
-                    x: 0.0,
-                    y: 0.0,
-                    width: 640.0,
-                    height: 480.0,
-                }),
-            ),
-        ];
-
-        for platform in [
-            MeetingPlatform::Zoom,
-            MeetingPlatform::GoogleMeet,
-            MeetingPlatform::MicrosoftTeams,
-            MeetingPlatform::Slack,
-            MeetingPlatform::Discord,
-            MeetingPlatform::Webex,
-        ] {
-            assert!(
-                find_participant_streams(&platform, &MeetingSurface::Native, &private_nodes)
-                    .is_empty(),
-                "unexpected participant for {platform:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_zoom_participant_anchor_is_platform_specific_across_surfaces() {
-        let participant = node(
-            22,
-            "AXGroup",
-            "Video render Ada Lovelace, Computer audio unmuted",
-            None,
-        );
-
-        assert!(
-            candidate_stream(
-                &MeetingPlatform::Zoom,
-                &MeetingSurface::Native,
-                &participant,
-            )
-            .is_some()
-        );
-        assert!(
-            candidate_stream(&MeetingPlatform::Zoom, &MeetingSurface::Web, &participant,).is_some()
-        );
-        for platform in [
-            MeetingPlatform::MicrosoftTeams,
-            MeetingPlatform::Discord,
-            MeetingPlatform::Webex,
-        ] {
-            assert!(candidate_stream(&platform, &MeetingSurface::Native, &participant).is_none());
-        }
-    }
-
-    #[test]
-    fn test_native_meeting_window_validation_is_evidence_backed() {
-        let settings = [
-            node(0, "AXWindow", "Zoom Workplace Settings", None),
-            node(1, "AXStaticText", "Video", None),
-            node(2, "AXButton", "Camera preview", None),
-        ];
-        let zoom_meeting = [node(
-            3,
-            "AXGroup",
-            "Video render Ada Lovelace, Computer audio unmuted",
-            None,
-        )];
-        let discord_voice = [node(4, "AXStaticText", "Voice connected", None)];
-
-        assert!(!native_meeting_window_is_validated(
-            &MeetingPlatform::Zoom,
-            &settings,
-        ));
-        assert!(native_meeting_window_is_validated(
-            &MeetingPlatform::Zoom,
-            &zoom_meeting,
-        ));
-        assert!(native_meeting_window_is_validated(
-            &MeetingPlatform::Discord,
-            &discord_voice,
-        ));
-        for platform in [MeetingPlatform::MicrosoftTeams, MeetingPlatform::Webex] {
-            assert!(!native_meeting_window_is_validated(
-                &platform,
-                &zoom_meeting,
-            ));
-        }
-        assert!(native_meeting_window_is_validated(
-            &MeetingPlatform::MicrosoftTeams,
-            &[fixture_node(5, "AXButton", "Hang up", &[0])],
-        ));
-        assert!(native_meeting_window_is_validated(
-            &MeetingPlatform::Webex,
-            &[fixture_node(6, "AXButton", "Leave meeting", &[0])],
-        ));
-    }
-
-    #[test]
-    fn test_unknown_browser_surface_does_not_emit_participant_streams() {
-        let streams = find_participant_streams(
-            &MeetingPlatform::Unknown,
-            &MeetingSurface::Unknown,
-            &[node(
-                8,
-                "AXGroup",
-                "Video render Private Browser Content, active speaker",
-                Some(AxRect {
-                    x: 0.0,
-                    y: 0.0,
-                    width: 320.0,
-                    height: 180.0,
-                }),
-            )],
-        );
-
-        assert!(streams.is_empty());
-    }
-
-    #[test]
-    fn test_meeting_chat_message_validation() {
-        assert!(validate_meeting_chat_message("disclosure message").is_ok());
-        assert!(validate_meeting_chat_message(" \n\t ").is_err());
-        assert!(validate_meeting_chat_message(&"x".repeat(2_000)).is_ok());
-        assert!(validate_meeting_chat_message(&"x".repeat(2_001)).is_err());
-    }
-
-    #[test]
-    fn test_chat_mutation_is_fail_closed_for_unvalidated_platforms() {
-        assert!(supports_meeting_chat_mutation("com.tinyspeck.slackmacgap"));
-        assert!(supports_meeting_chat_mutation("com.slack.Slack"));
-        for bundle_id in [
-            "us.zoom.xos",
-            "com.microsoft.teams2",
-            "com.hnc.Discord",
-            "Cisco-Systems.Spark",
-            "com.google.Chrome",
-        ] {
-            assert!(!supports_meeting_chat_mutation(bundle_id));
-        }
-    }
-
-    #[test]
-    fn test_established_native_bundle_aliases_are_classified() {
-        for (bundle_id, platform) in [
-            ("com.slack.Slack", MeetingPlatform::Slack),
-            ("com.cisco.webex", MeetingPlatform::Webex),
-            ("com.cisco.webexmeetingsapp", MeetingPlatform::Webex),
-            ("com.discordapp.Discord", MeetingPlatform::Discord),
-        ] {
-            assert!(is_meeting_app_bundle(bundle_id));
-            assert_eq!(classify_bundle(bundle_id), platform);
-            assert_eq!(
-                classify_surface(bundle_id, &platform),
-                MeetingSurface::Native
-            );
-        }
-    }
-
-    #[test]
-    fn test_established_browser_variants_are_recognized_as_web_surfaces() {
-        for bundle_id in [
-            "com.apple.SafariTechnologyPreview",
-            "com.google.Chrome.canary",
-            "com.microsoft.edgemac.Beta",
-            "com.microsoft.edgemac.Canary",
-            "com.microsoft.edgemac.Dev",
-            "org.mozilla.firefoxdeveloperedition",
-            "org.mozilla.nightly",
-            "com.brave.Browser.beta",
-            "com.brave.Browser.nightly",
-            "org.chromium.Chromium",
-            "com.operasoftware.OperaDeveloper",
-            "com.operasoftware.OperaGX",
-            "com.operasoftware.OperaNext",
-            "net.imput.helium",
-        ] {
-            assert!(is_meeting_app_bundle(bundle_id));
-            assert!(is_browser_bundle(bundle_id));
-            assert_eq!(
-                classify_surface(bundle_id, &MeetingPlatform::Unknown),
-                MeetingSurface::Web
-            );
-        }
-    }
-
-    #[test]
-    fn test_meeting_app_registry_drives_bundle_kind() {
-        let mut seen = HashSet::new();
-
-        for bundle in MEETING_APP_BUNDLES {
-            assert!(seen.insert(bundle.id), "duplicate bundle id: {}", bundle.id);
-            assert!(is_meeting_app_bundle(bundle.id));
-            assert_eq!(
-                is_browser_bundle(bundle.id),
-                bundle.kind == MeetingAppBundleKind::Browser,
-                "unexpected browser classification for {}",
-                bundle.id
-            );
-        }
-    }
-
-    #[test]
-    fn test_chat_mutation_scope_deduplicates_one_recognized_meeting_app() {
-        let bundle_ids = vec![
-            "com.tinyspeck.slackmacgap".to_string(),
-            "com.tinyspeck.slackmacgap".to_string(),
-            "com.hyprnote.dev".to_string(),
-        ];
-
-        assert_eq!(
-            unique_recognized_meeting_bundle(&bundle_ids),
-            Ok("com.tinyspeck.slackmacgap")
-        );
-    }
-
-    #[test]
-    fn test_chat_mutation_scope_rejects_zero_or_multiple_meeting_apps() {
-        assert!(unique_recognized_meeting_bundle(&[]).is_err());
-        assert!(
-            unique_recognized_meeting_bundle(&[
-                "us.zoom.xos".to_string(),
-                "com.tinyspeck.slackmacgap".to_string(),
-            ])
-            .is_err()
-        );
-    }
-
-    #[test]
-    fn test_zoom_scope_does_not_fall_back_to_an_unrelated_slack_huddle() {
-        let bundle_ids = ["us.zoom.xos".to_string()];
-        let scoped_bundle = unique_recognized_meeting_bundle(&bundle_ids).unwrap();
-
-        assert_eq!(scoped_bundle, "us.zoom.xos");
-        assert!(!supports_meeting_chat_mutation(scoped_bundle));
-    }
-
-    #[test]
-    fn test_chat_input_candidate_requires_chat_signal() {
-        let mut chat = node(3, "AXTextArea", "Send a message", None);
-        chat.settable_value = true;
-        chat.text = node_text(
-            &chat.role,
-            &chat.title,
-            &chat.value,
-            &chat.description,
-            &chat.placeholder,
-        );
-
-        let target = candidate_chat_target(&chat).unwrap();
-
-        assert_eq!(target.kind, "input");
-        assert!(target.settable);
-        assert!(target.confidence > 0.7);
-    }
-
-    #[test]
-    fn test_visible_empty_chat_surface_can_establish_a_capture_baseline() {
-        let mut zoom_chat_list = node(4, "AXTable", "Chat list", None);
-        zoom_chat_list.within_zoom_meeting_scope = true;
-        zoom_chat_list.within_zoom_chat_scope = true;
-        assert!(meeting_chat_surface_is_visible(
-            &MeetingPlatform::Zoom,
-            &[zoom_chat_list],
-        ));
-
-        let mut zoom_rename_input = node(5, "AXTextField", "Display name", None);
-        zoom_rename_input.settable_value = true;
-        zoom_rename_input.within_zoom_meeting_scope = true;
-        zoom_rename_input.text = node_text(
-            &zoom_rename_input.role,
-            &zoom_rename_input.title,
-            &zoom_rename_input.value,
-            &zoom_rename_input.description,
-            &zoom_rename_input.placeholder,
-        );
-        assert!(!meeting_chat_surface_is_visible(
-            &MeetingPlatform::Zoom,
-            &[zoom_rename_input],
-        ));
-
-        let mut slack_input = node(6, "AXTextArea", "Message to test", None);
-        slack_input.settable_value = true;
-        slack_input.within_slack_huddle_scope = true;
-        slack_input.text = node_text(
-            &slack_input.role,
-            &slack_input.title,
-            &slack_input.value,
-            &slack_input.description,
-            &slack_input.placeholder,
-        );
-        assert!(meeting_chat_surface_is_visible(
-            &MeetingPlatform::Slack,
-            &[slack_input],
-        ));
-
-        assert!(!meeting_chat_surface_is_visible(
-            &MeetingPlatform::Slack,
-            &[node(7, "AXTextArea", "Search", None)],
-        ));
-    }
-
-    #[test]
-    fn test_browser_chat_scope_requires_live_exit_visible_composer_and_platform_container() {
-        let meet_nodes = vec![
-            fixture_node(0, "AXWebArea", "Team sync - Google Meet", &[]),
-            fixture_node(1, "AXButton", "Leave call", &[0]),
-            fixture_node(2, "AXGroup", "In-call messages", &[1]),
-            fixture_composer(3, "Send a message", &[1, 0]),
-        ];
-
-        assert_eq!(
-            validated_chat_scope(&MeetingPlatform::GoogleMeet, &meet_nodes),
-            Some((vec![1], vec![1, 0]))
-        );
-
-        let mut prejoin_nodes = meet_nodes.clone();
-        prejoin_nodes[1] = fixture_node(1, "AXButton", "Turn off microphone", &[0]);
-        assert!(validated_chat_scope(&MeetingPlatform::GoogleMeet, &prejoin_nodes).is_none());
-
-        let mut hidden_composer_nodes = meet_nodes.clone();
-        hidden_composer_nodes[3].bounds = None;
-        assert!(
-            validated_chat_scope(&MeetingPlatform::GoogleMeet, &hidden_composer_nodes).is_none()
-        );
-
-        let mut duplicate_composer_nodes = meet_nodes.clone();
-        duplicate_composer_nodes.push(fixture_composer(4, "Send a message", &[1, 1]));
-        assert!(
-            validated_chat_scope(&MeetingPlatform::GoogleMeet, &duplicate_composer_nodes).is_none()
-        );
-
-        let mut support_widget_nodes = meet_nodes.clone();
-        support_widget_nodes[2] = fixture_node(2, "AXGroup", "Support chat", &[1]);
-        assert!(
-            validated_chat_scope(&MeetingPlatform::GoogleMeet, &support_widget_nodes).is_none()
-        );
-    }
-
-    #[test]
-    fn test_platform_chat_adapters_validate_the_requested_provider_matrix() {
-        for (platform, exit_label, scope_label, composer_label) in [
-            (
-                MeetingPlatform::GoogleMeet,
-                "Leave call",
-                "In-call messages",
-                "Send a message",
-            ),
-            (
-                MeetingPlatform::MicrosoftTeams,
-                "Hang up",
-                "Meeting chat",
-                "Type a message",
-            ),
-            (
-                MeetingPlatform::Zoom,
-                "Leave meeting",
-                "Chat",
-                "Message everyone",
-            ),
-            (
-                MeetingPlatform::Webex,
-                "Leave meeting",
-                "Chat with everyone",
-                "Type a message",
-            ),
-        ] {
-            let nodes = vec![
-                fixture_node(0, "AXWebArea", "Meeting", &[]),
-                fixture_node(1, "AXButton", exit_label, &[0]),
-                fixture_node(2, "AXGroup", scope_label, &[1]),
-                fixture_composer(3, composer_label, &[1, 0]),
-            ];
-
-            assert_eq!(
-                validated_chat_scope(&platform, &nodes),
-                Some((vec![1], vec![1, 0])),
-                "adapter did not validate {platform:?}"
-            );
-        }
-
-        let slack_nodes = vec![
-            fixture_node(0, "AXWebArea", "Huddle in test", &[]),
-            fixture_node(1, "AXButton", "Leave huddle", &[0]),
-            fixture_node(2, "AXGroup", "Thread in test", &[1]),
-            fixture_composer(3, "Message to test", &[1, 0]),
-        ];
-        assert_eq!(
-            validated_chat_scope(&MeetingPlatform::Slack, &slack_nodes),
-            Some((vec![1], vec![1, 0]))
-        );
-
-        let mut ordinary_channel_nodes = slack_nodes;
-        ordinary_channel_nodes[2] = fixture_node(2, "AXGroup", "Channel test", &[1]);
-        assert!(validated_chat_scope(&MeetingPlatform::Slack, &ordinary_channel_nodes).is_none());
-    }
-
-    #[test]
-    fn test_teams_and_webex_reject_generic_chat_containers() {
-        for (platform, exit_label, composer_label) in [
-            (MeetingPlatform::MicrosoftTeams, "Hang up", "Type a message"),
-            (MeetingPlatform::Webex, "Leave meeting", "Send a message"),
-        ] {
-            let nodes = vec![
-                fixture_node(0, "AXWebArea", "Meeting", &[]),
-                fixture_node(1, "AXButton", exit_label, &[0]),
-                fixture_node(2, "AXGroup", "Chat", &[1]),
-                fixture_composer(3, composer_label, &[1, 0]),
-            ];
-
-            assert!(
-                validated_chat_scope(&platform, &nodes).is_none(),
-                "generic chat unexpectedly validated for {platform:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_web_capture_extracts_only_the_validated_chat_subtree() {
-        let nodes = vec![
-            fixture_node(0, "AXWebArea", "Team sync - Google Meet", &[]),
-            fixture_node(1, "AXButton", "Leave call", &[0]),
-            fixture_node(2, "AXGroup", "In-call messages", &[1]),
-            fixture_composer(3, "Send a message", &[1, 0]),
-            fixture_node(
-                4,
-                "AXGroup",
-                "Ada Lovelace\n10:42 AM\nDiscuss the rollout https://example.com/plan",
-                &[1, 1],
-            ),
-            fixture_node(
-                5,
-                "AXGroup",
-                "Mallory\n10:43 AM\nUnrelated browser content",
-                &[2, 0],
-            ),
-        ];
-
-        let messages =
-            extract_chat_messages(&MeetingPlatform::GoogleMeet, &MeetingSurface::Web, &nodes);
-
-        assert_eq!(messages.len(), 1);
-        assert_eq!(messages[0].sender.as_deref(), Some("Ada Lovelace"));
-        assert_eq!(messages[0].timestamp.as_deref(), Some("10:42 AM"));
-        assert_eq!(
-            messages[0].text,
-            "Discuss the rollout https://example.com/plan"
-        );
-        assert_eq!(messages[0].links, vec!["https://example.com/plan"]);
-    }
-
-    #[test]
-    fn test_web_capture_ignores_aggregate_containers() {
-        let mut aggregate_list = fixture_node(4, "AXList", "Message list", &[1, 1]);
-        aggregate_list.value = Some("Mallory\n10:41 AM\nAggregated list value".to_string());
-        let nodes = vec![
-            fixture_node(0, "AXWebArea", "Team sync - Google Meet", &[]),
-            fixture_node(1, "AXButton", "Leave call", &[0]),
-            fixture_node(2, "AXGroup", "In-call messages", &[1]),
-            fixture_composer(3, "Send a message", &[1, 0]),
-            aggregate_list,
-            fixture_node(
-                5,
-                "AXGroup",
-                "Ada Lovelace\n10:42 AM\nFirst message\nGrace Hopper\n10:43 AM\nSecond message",
-                &[1, 2],
-            ),
-            fixture_node(
-                6,
-                "AXGroup",
-                "Ada Lovelace\n10:42 AM\nFirst message",
-                &[1, 2, 0],
-            ),
-            fixture_node(
-                7,
-                "AXGroup",
-                "Grace Hopper\n10:43 AM\nSecond message",
-                &[1, 2, 1],
-            ),
-        ];
-
-        let messages =
-            extract_chat_messages(&MeetingPlatform::GoogleMeet, &MeetingSurface::Web, &nodes);
-
-        assert_eq!(messages.len(), 2);
-        assert_eq!(messages[0].text, "First message");
-        assert_eq!(messages[1].text, "Second message");
-    }
-
-    #[test]
-    fn test_browser_context_preserves_query_identified_meeting_identity() {
-        let teams_nodes = vec![
-            fixture_node(0, "AXWebArea", "Microsoft Teams", &[]),
-            fixture_node(1, "AXButton", "Hang up", &[0]),
-            fixture_node(2, "AXGroup", "Meeting chat", &[1]),
-            fixture_composer(3, "Type a message", &[1, 0]),
-        ];
-        let root = |url: &str| BrowserMeetingRoot {
-            platform: MeetingPlatform::MicrosoftTeams,
-            window_title: Some("Microsoft Teams meeting".to_string()),
-            web_area_url: Some(url.to_string()),
-            nodes: teams_nodes.clone(),
-        };
-
-        let first = browser_capture_context_id(&root(
-            "https://teams.microsoft.com/v2/?meetingId=first#fragment",
-        ))
-        .unwrap();
-        let same_without_fragment =
-            browser_capture_context_id(&root("https://teams.microsoft.com/v2/?meetingId=first"))
-                .unwrap();
-        let second =
-            browser_capture_context_id(&root("https://teams.microsoft.com/v2/?meetingId=second"))
-                .unwrap();
-
-        assert_eq!(first, same_without_fragment);
-        assert_ne!(first, second);
-    }
-
-    #[test]
-    fn test_capture_context_ignores_volatile_titles_and_tree_paths() {
-        let nodes = |scope_path: &[usize], composer_path: &[usize], root_role: &str| {
-            vec![
-                fixture_node(0, root_role, "Microsoft Teams", &[]),
-                fixture_node(1, "AXButton", "Hang up", &[0]),
-                fixture_node(2, "AXGroup", "Meeting chat", scope_path),
-                fixture_composer(3, "Type a message", composer_path),
-            ]
-        };
-        let browser_root =
-            |title: &str, scope_path: &[usize], composer_path: &[usize]| BrowserMeetingRoot {
-                platform: MeetingPlatform::MicrosoftTeams,
-                window_title: Some(title.to_string()),
-                web_area_url: Some("https://teams.microsoft.com/v2/?meetingId=stable".to_string()),
-                nodes: nodes(scope_path, composer_path, "AXWebArea"),
-            };
-
-        let first =
-            browser_capture_context_id(&browser_root("Microsoft Teams meeting", &[1], &[1, 0]))
-                .unwrap();
-        let shifted =
-            browser_capture_context_id(&browser_root("(1) Microsoft Teams meeting", &[5], &[5, 2]))
-                .unwrap();
-
-        assert_eq!(first, shifted);
-
-        let first_native = NativeMeetingRoot {
-            window_title: Some("Microsoft Teams meeting".to_string()),
-            nodes: nodes(&[1], &[1, 0], "AXWindow"),
-        };
-        let shifted_native = NativeMeetingRoot {
-            window_title: Some("Microsoft Teams meeting · new activity".to_string()),
-            nodes: nodes(&[5], &[5, 2], "AXWindow"),
-        };
-
-        assert_eq!(
-            native_capture_context_id(&MeetingPlatform::MicrosoftTeams, &first_native),
-            native_capture_context_id(&MeetingPlatform::MicrosoftTeams, &shifted_native)
-        );
-    }
-
-    #[test]
-    fn test_chat_button_is_open_chat_control_not_input() {
-        let target = candidate_chat_target(&node(4, "AXButton", "Chat", None)).unwrap();
-
-        assert_eq!(target.kind, "openChatControl");
-        assert!(!target.settable);
-        assert!(target.signals.contains(&"open-chat-control".to_string()));
-    }
-
-    #[test]
-    fn test_zoom_chat_message_parser_preserves_sender_time_text_and_links() {
-        let parsed = parse_chat_message(
-            &MeetingPlatform::Zoom,
-            "From Ada Lovelace to Everyone\n10:42 AM\nHere is the doc https://example.com/spec.",
-        )
-        .unwrap();
-
-        assert_eq!(parsed.sender, Some("Ada Lovelace".to_string()));
-        assert_eq!(parsed.timestamp, Some("10:42 AM".to_string()));
-        assert_eq!(parsed.text, "Here is the doc https://example.com/spec.");
-        assert_eq!(
-            extract_links(&parsed.text),
-            vec!["https://example.com/spec"]
-        );
-    }
-
-    #[test]
-    fn test_zoom_chat_message_parser_handles_current_native_row_description() {
-        let parsed = parse_chat_message(
-            &MeetingPlatform::Zoom,
-            "You, ANLG-76 AX integration 5844 https://example.com/ax-test, 4:16\u{202f}PM",
-        )
-        .unwrap();
-
-        assert_eq!(parsed.sender, Some("You".to_string()));
-        assert_eq!(parsed.timestamp, Some("4:16 PM".to_string()));
-        assert_eq!(
-            parsed.text,
-            "ANLG-76 AX integration 5844 https://example.com/ax-test"
-        );
-        assert_eq!(
-            extract_links(&parsed.text),
-            vec!["https://example.com/ax-test"]
-        );
-    }
-
-    #[test]
-    fn test_zoom_chat_direction_uses_native_self_sender_label() {
-        assert_eq!(
-            meeting_chat_direction(&MeetingPlatform::Zoom, Some("You")),
-            Some(MeetingChatDirection::Outgoing)
-        );
-        assert_eq!(
-            meeting_chat_direction(&MeetingPlatform::Zoom, Some("Ada")),
-            Some(MeetingChatDirection::Incoming)
-        );
-        assert_eq!(
-            meeting_chat_direction(&MeetingPlatform::Slack, Some("You")),
-            None
-        );
-    }
-
-    #[test]
-    fn test_zoom_capture_requires_zoom_meeting_window_scope() {
-        assert!(is_zoom_meeting_scope_node(&node(
-            0,
-            "AXWindow",
-            "Zoom Meeting",
-            None,
-        )));
-        assert!(is_zoom_meeting_scope_node(&node(
-            1,
-            "AXWindow",
-            "John Jeong's Zoom Meeting",
-            None,
-        )));
-        assert!(!is_zoom_meeting_scope_node(&node(
-            2,
-            "AXWindow",
-            "Zoom Workplace",
-            None,
-        )));
-
-        let mut chat_row = node(3, "AXGroup", "You, meeting chat message, 4:16 PM", None);
-        chat_row.identifier = Some("ZMTextMessageCellView".to_string());
-        assert!(is_zoom_chat_scope_node(&chat_row));
-
-        let mut meeting_caption = node(
-            4,
-            "AXStaticText",
-            "Ada, confidential caption, 4:16 PM",
-            None,
-        );
-        meeting_caption.within_zoom_meeting_scope = true;
-        assert!(
-            extract_chat_messages(
-                &MeetingPlatform::Zoom,
-                &MeetingSurface::Native,
-                &[meeting_caption],
-            )
-            .is_empty()
-        );
-
-        let team_chat_message = node(3, "AXStaticText", "You, private team chat, 4:16 PM", None);
-        assert!(
-            extract_chat_messages(
-                &MeetingPlatform::Zoom,
-                &MeetingSurface::Native,
-                &[team_chat_message],
-            )
-            .is_empty()
-        );
-    }
-
-    #[test]
-    fn test_slack_chat_message_parser_handles_sender_time_prefix() {
-        let parsed = parse_chat_message(
-            &MeetingPlatform::Slack,
-            "Grace Hopper 9:03 PM\nShip it after the final check",
-        )
-        .unwrap();
-
-        assert_eq!(parsed.sender, Some("Grace Hopper".to_string()));
-        assert_eq!(parsed.timestamp, Some("9:03 PM".to_string()));
-        assert_eq!(parsed.text, "Ship it after the final check");
-    }
-
-    #[test]
-    fn test_slack_chat_message_parser_handles_native_accessibility_description() {
-        let parsed = parse_chat_message(
-            &MeetingPlatform::Slack,
-            "John Jeong: @Artem lorem ipsum. Friday at 5:50\u{202f}PM.",
-        )
-        .unwrap();
-
-        assert_eq!(parsed.sender, Some("John Jeong".to_string()));
-        assert_eq!(parsed.timestamp, Some("5:50 PM".to_string()));
-        assert_eq!(parsed.text, "@Artem lorem ipsum");
-    }
-
-    #[test]
-    fn test_web_chat_parsers_cover_meet_teams_zoom_slack_and_webex_shapes() {
-        for (platform, raw_text, sender, timestamp, text) in [
-            (
-                MeetingPlatform::GoogleMeet,
-                "Ada Lovelace\n10:42 AM\nMeet decision",
-                "Ada Lovelace",
-                "10:42 AM",
-                "Meet decision",
-            ),
-            (
-                MeetingPlatform::MicrosoftTeams,
-                "Grace Hopper 10:43 AM\nTeams decision",
-                "Grace Hopper",
-                "10:43 AM",
-                "Teams decision",
-            ),
-            (
-                MeetingPlatform::Zoom,
-                "Linus Torvalds, Zoom decision, 10:44 AM",
-                "Linus Torvalds",
-                "10:44 AM",
-                "Zoom decision",
-            ),
-            (
-                MeetingPlatform::Slack,
-                "Margaret Hamilton\nSlack decision\n10:45 AM",
-                "Margaret Hamilton",
-                "10:45 AM",
-                "Slack decision",
-            ),
-            (
-                MeetingPlatform::Webex,
-                "Katherine Johnson\n10:46 AM\nWebex decision",
-                "Katherine Johnson",
-                "10:46 AM",
-                "Webex decision",
-            ),
-        ] {
-            let parsed = parse_chat_message(&platform, raw_text)
-                .unwrap_or_else(|| panic!("failed to parse {platform:?}"));
-            assert_eq!(parsed.sender.as_deref(), Some(sender));
-            assert_eq!(parsed.timestamp.as_deref(), Some(timestamp));
-            assert_eq!(parsed.text, text);
-        }
-    }
-
-    #[test]
-    fn test_web_speaker_mapping_requires_an_explicit_accessibility_state() {
-        for platform in [
-            MeetingPlatform::Zoom,
-            MeetingPlatform::GoogleMeet,
-            MeetingPlatform::MicrosoftTeams,
-            MeetingPlatform::Slack,
-            MeetingPlatform::Webex,
-        ] {
-            let speaker = fixture_node(40, "AXGroup", "Ada Lovelace, active speaker", &[4, 0]);
-            let stream = candidate_stream(&platform, &MeetingSurface::Web, &speaker)
-                .unwrap_or_else(|| panic!("failed to map {platform:?} speaker"));
-            assert_eq!(stream.participant_name.as_deref(), Some("Ada Lovelace"));
-            assert!(stream.is_active_speaker);
-
-            let chat_text = fixture_node(
-                41,
-                "AXStaticText",
-                "We should discuss active speaker behavior",
-                &[4, 1],
-            );
-            assert!(candidate_stream(&platform, &MeetingSurface::Web, &chat_text).is_none());
-        }
-
-        for label in [
-            "Ada Lovelace, speaking French",
-            "Ada Lovelace, speaking=false",
-            "Ada Lovelace, active speaker=false",
-            "Ada Lovelace, active speaker (false)",
-            "Ada Lovelace, active speaker, false",
-            "Ada Lovelace, speaking, French",
-            "Active speaker: false",
-            "who is speaking",
-            "We should discuss who is speaking",
-            "someone is speaking",
-        ] {
-            let false_state = fixture_node(42, "AXGroup", label, &[4, 2]);
-            assert!(
-                candidate_stream(
-                    &MeetingPlatform::GoogleMeet,
-                    &MeetingSurface::Web,
-                    &false_state,
-                )
-                .is_none()
-            );
-        }
-
-        let false_zoom_state = fixture_node(
-            43,
-            "AXGroup",
-            "Video render Ada Lovelace, speaking=false",
-            &[4, 3],
-        );
-        assert!(
-            candidate_stream(
-                &MeetingPlatform::Zoom,
-                &MeetingSurface::Native,
-                &false_zoom_state,
-            )
-            .is_none()
-        );
-    }
-
-    #[test]
-    fn test_zoom_participant_names_cover_web_speaker_and_native_roster_labels() {
-        assert_eq!(
-            participant_name_from_evidence(
-                &MeetingPlatform::Zoom,
-                "Video render Grace Hopper is speaking",
-            )
-            .as_deref(),
-            Some("Grace Hopper")
-        );
-        assert_eq!(
-            participant_name_from_evidence(
-                &MeetingPlatform::Zoom,
-                "Video render Grace Hopper, Computer audio unmuted, active speaker",
-            )
-            .as_deref(),
-            Some("Grace Hopper")
-        );
-        assert_eq!(
-            participant_name_from_evidence(&MeetingPlatform::Zoom, "Ada Lovelace, active speaker",)
-                .as_deref(),
-            Some("Ada Lovelace")
-        );
-        assert_eq!(
-            participant_name_from_evidence(
-                &MeetingPlatform::Zoom,
-                "Ada Lovelace (Host, me, Participant ID:417329) No audio connected",
-            )
-            .as_deref(),
-            Some("Ada Lovelace")
-        );
-    }
-
-    #[test]
-    fn test_past_slack_huddle_thread_is_not_captured_without_active_huddle() {
-        let mut message = node(
-            0,
-            "AXGroup",
-            "John Jeong: @Artem lorem ipsum. Friday at 5:50 PM.",
-            None,
-        );
-        message.within_slack_huddle_scope = true;
-
-        assert!(
-            extract_chat_messages(&MeetingPlatform::Slack, &MeetingSurface::Native, &[message],)
-                .is_empty()
-        );
-    }
-
-    #[test]
-    fn test_chat_parsers_reject_unstructured_static_text() {
-        assert!(
-            parse_chat_message(
-                &MeetingPlatform::Zoom,
-                "Recording has started for this meeting"
-            )
-            .is_none()
-        );
-        assert!(parse_chat_message(&MeetingPlatform::Slack, "Channels\nGeneral").is_none());
-    }
-
-    #[test]
-    fn test_chat_parsers_reject_invalid_timestamps() {
-        assert!(!looks_like_time("99:99"));
-        assert!(!looks_like_time("13:00 PM"));
-        assert!(looks_like_time("23:59"));
-        assert!(looks_like_time("12:59 PM"));
-    }
-
-    #[test]
-    fn test_active_bundle_selection_is_scoped_and_deduplicated() {
-        let active_bundle_ids = vec![
-            "com.tinyspeck.slackmacgap".to_string(),
-            "com.google.Chrome".to_string(),
-            "com.tinyspeck.slackmacgap".to_string(),
-            "com.example.unrelated".to_string(),
-        ];
-
-        assert_eq!(
-            select_active_bundle_ids(
-                MEETING_APP_BUNDLES.iter().map(|bundle| bundle.id),
-                &active_bundle_ids,
-            ),
-            vec!["com.tinyspeck.slackmacgap", "com.google.Chrome"]
-        );
-        assert!(
-            select_active_bundle_ids(MEETING_APP_BUNDLES.iter().map(|bundle| bundle.id), &[],)
-                .is_empty()
-        );
-    }
-
-    #[test]
-    fn test_slack_capture_requires_active_huddle_and_huddle_specific_scope() {
-        let active_control = node(0, "AXButton", "Leave huddle", None);
-        let channel_message = node(
-            1,
-            "AXStaticText",
-            "Grace Hopper 9:03 PM\nChannel-only message",
-            None,
-        );
-
-        assert!(
-            extract_chat_messages(
-                &MeetingPlatform::Slack,
-                &MeetingSurface::Native,
-                &[active_control.clone(), channel_message.clone()],
-            )
-            .is_empty()
-        );
-
-        let mut huddle_message = channel_message;
-        huddle_message.within_slack_huddle_scope = true;
-        assert!(
-            extract_chat_messages(
-                &MeetingPlatform::Slack,
-                &MeetingSurface::Native,
-                &[huddle_message.clone()],
-            )
-            .is_empty()
-        );
-
-        let messages = extract_chat_messages(
-            &MeetingPlatform::Slack,
-            &MeetingSurface::Native,
-            &[active_control, huddle_message],
-        );
-        assert_eq!(messages.len(), 1);
-        assert_eq!(messages[0].text, "Channel-only message");
-    }
-
-    #[test]
-    fn test_slack_huddle_scope_rejects_general_channel_containers() {
-        assert!(is_slack_huddle_scope_node(&node(
-            0,
-            "AXGroup",
-            "Huddle chat",
-            None,
-        )));
-        assert!(!is_slack_huddle_scope_node(&node(
-            1,
-            "AXGroup",
-            "#general conversation",
-            None,
-        )));
-        assert!(!slack_huddle_is_active(&[node(
-            2,
-            "AXStaticText",
-            "Someone mentioned leave huddle in a channel message",
-            None,
-        )]));
-    }
-
-    #[test]
-    fn test_slack_capture_context_identity_tracks_validated_surface() {
-        let context = slack_capture_context_id("test", "Huddle in test", 0x101, 0x202);
-
-        assert_eq!(
-            context,
-            slack_capture_context_id("test", "Huddle in test", 0x101, 0x202)
-        );
-        assert_ne!(
-            context,
-            slack_capture_context_id("another", "Huddle in another", 0x101, 0x202)
-        );
-        assert_ne!(
-            context,
-            slack_capture_context_id("test", "Huddle in test", 0x303, 0x202)
-        );
-        assert_ne!(
-            context,
-            slack_capture_context_id("test", "Huddle in test", 0x101, 0x404)
-        );
-    }
-
-    #[test]
-    fn test_zoom_capture_context_stays_stable_across_participant_changes() {
-        let root = |chat_hash, participant_names: &[&str]| {
-            let mut window = node(0, "AXWindow", "Zoom Meeting", None);
-            window.element_hash = Some(0x101);
-            window.within_zoom_meeting_scope = true;
-
-            let mut chat = node(1, "AXTable", "Chat list", None);
-            chat.element_hash = Some(chat_hash);
-            chat.within_zoom_meeting_scope = true;
-            chat.within_zoom_chat_scope = true;
-
-            let participants = participant_names.iter().enumerate().map(|(index, name)| {
-                let mut participant = node(
-                    index + 2,
-                    "AXGroup",
-                    &format!("Video render {name}, Computer audio unmuted"),
-                    None,
-                );
-                participant.element_hash = Some(0x300 + index);
-                participant.within_zoom_meeting_scope = true;
-                participant
-            });
-
-            NativeMeetingRoot {
-                window_title: Some("Zoom Meeting".to_string()),
-                nodes: std::iter::once(window)
-                    .chain(std::iter::once(chat))
-                    .chain(participants)
-                    .collect(),
-            }
-        };
-
-        let first = zoom_capture_context_id(&root(0x202, &["Ada", "Grace"])).unwrap();
-        let reordered = zoom_capture_context_id(&root(0x202, &["Grace", "Ada"])).unwrap();
-        let switched = zoom_capture_context_id(&root(0x202, &["Ada", "Linus"])).unwrap();
-        let new_chat_surface = zoom_capture_context_id(&root(0x404, &["Ada", "Grace"])).unwrap();
-
-        assert_eq!(first, reordered);
-        assert_eq!(first, switched);
-        assert_ne!(first, new_chat_surface);
-    }
-
-    #[test]
-    fn test_extract_chat_messages_keeps_repeated_source_rows_distinct() {
-        let message = "From Ada Lovelace to Everyone\n10:42 AM\nDecision: keep the launch date";
-        let nodes = vec![
-            zoom_message_node(12, message),
-            zoom_message_node(13, message),
-        ];
-
-        let messages =
-            extract_chat_messages(&MeetingPlatform::Zoom, &MeetingSurface::Native, &nodes);
-
-        assert_eq!(messages.len(), 2);
-        assert_eq!(messages[0].sender, Some("Ada Lovelace".to_string()));
-        assert_eq!(messages[0].text, "Decision: keep the launch date");
-        assert_ne!(messages[0].id, messages[1].id);
-        assert!(messages[0].id.ends_with("occurrence=1"));
-        assert!(messages[1].id.ends_with("occurrence=2"));
-    }
-
-    #[test]
-    fn test_element_hash_stabilizes_identical_rows_across_snapshot_shifts() {
-        let message = "From Ada Lovelace to Everyone\n10:42 AM\nDecision: keep the launch date";
-        let hashed_node = |index, element_hash| {
-            let mut node = zoom_message_node(index, message);
-            node.element_hash = Some(element_hash);
-            node
-        };
-
-        let first = extract_chat_messages(
-            &MeetingPlatform::Zoom,
-            &MeetingSurface::Native,
-            &[hashed_node(0, 0x101), hashed_node(1, 0x202)],
-        );
-        let shifted = extract_chat_messages(
-            &MeetingPlatform::Zoom,
-            &MeetingSurface::Native,
-            &[
-                hashed_node(0, 0x303),
-                hashed_node(1, 0x101),
-                hashed_node(2, 0x202),
-            ],
-        );
-
-        assert_eq!(first[0].id, shifted[1].id);
-        assert_eq!(first[1].id, shifted[2].id);
-        assert!(first[0].id.contains("cfhash=101"));
-        assert!(first[0].id.contains("Decision: keep the launch date"));
-    }
-
-    #[test]
-    fn test_extract_chat_messages_retains_newest_eighty_rows() {
-        let nodes = (0..85)
-            .map(|index| {
-                zoom_message_node(index, &format!("From Ada to Everyone\nmessage {index}"))
-            })
-            .collect::<Vec<_>>();
-
-        let messages =
-            extract_chat_messages(&MeetingPlatform::Zoom, &MeetingSurface::Native, &nodes);
-
-        assert_eq!(messages.len(), 80);
-        assert_eq!(messages.first().unwrap().text, "message 5");
-        assert_eq!(messages.last().unwrap().text, "message 84");
-    }
-
-    #[test]
-    fn test_zoom_participant_roster_row_becomes_stream_candidate() {
-        let streams = find_participant_streams(
-            &MeetingPlatform::Zoom,
-            &MeetingSurface::Native,
-            &[node(
-                11,
-                "AXStaticText",
-                "Ada Lovelace (Host, me, Participant ID:417329) No audio connected",
-                None,
-            )],
-        );
-
-        assert_eq!(streams.len(), 1);
-        assert!(
-            streams[0]
-                .signals
-                .contains(&"participant-row-label".to_string())
-        );
-        assert!(
-            streams[0]
-                .signals
-                .contains(&"audio-state-label".to_string())
-        );
-    }
-
-    #[test]
-    fn test_browser_title_classifies_meet_web() {
-        let web_area = node(16, "AXWebArea", "Team sync - Google Meet", None);
-        assert_eq!(
-            classify_browser_context(
-                Some("https://meet.google.com/abc-defg-hij"),
-                Some("Team sync - Google Meet - Google Chrome"),
-                Some(&web_area),
-                &[],
-            ),
-            MeetingPlatform::GoogleMeet
-        );
-        assert_eq!(
-            classify_surface("com.google.Chrome", &MeetingPlatform::GoogleMeet),
-            MeetingSurface::Web
-        );
-    }
-
-    #[test]
-    fn test_browser_background_tab_nodes_cannot_classify_active_window() {
-        let background_meet_node = node(
-            17,
-            "AXButton",
-            "Team sync - Google Meet background tab",
-            None,
-        );
-
-        assert_eq!(
-            classify_platform(
-                "com.google.Chrome",
-                Some("Inbox - Google Chrome"),
-                &[background_meet_node],
-                MeetingPlatform::Unknown,
-            ),
-            MeetingPlatform::Unknown
-        );
-    }
-
-    #[test]
-    fn test_browser_active_web_area_can_validate_one_platform_but_not_conflicts() {
-        let meet_web_area = node(18, "AXWebArea", "Team sync - Google Meet", None);
-        let generic_web_area = node(19, "AXWebArea", "Document", None);
-        assert_eq!(
-            classify_browser_context(
-                Some("https://meet.google.com/abc-defg-hij"),
-                Some("Google Chrome"),
-                Some(&meet_web_area),
-                &[],
-            ),
-            MeetingPlatform::GoogleMeet
-        );
-
-        assert_eq!(
-            classify_browser_context(
-                Some("https://meet.google.com/abc-defg-hij"),
-                Some("Zoom Meeting - Google Chrome"),
-                Some(&meet_web_area),
-                &[],
-            ),
-            MeetingPlatform::Unknown
-        );
-
-        assert_eq!(
-            classify_browser_context(
-                Some("https://meet.google.com/abc-defg-hij"),
-                Some("Google Chrome"),
-                Some(&generic_web_area),
-                &[],
-            ),
-            MeetingPlatform::Unknown
-        );
-        assert_eq!(
-            classify_browser_context(
-                Some("https://meet.google.com/abc-defg-hij"),
-                Some("Google Chrome"),
-                Some(&generic_web_area),
-                &[node(20, "AXButton", "Leave call", None)],
-            ),
-            MeetingPlatform::GoogleMeet
-        );
-
-        assert_eq!(
-            classify_browser_context(
-                Some("https://www.google.com/search?q=Google+Meet"),
-                Some("Google Meet - Google Search"),
-                Some(&meet_web_area),
-                &[],
-            ),
-            MeetingPlatform::Unknown
-        );
-    }
-
-    #[test]
-    fn test_browser_meeting_origins_are_exact_and_https_only() {
-        for (url, platform) in [
-            (
-                "https://meet.google.com/abc-defg-hij",
-                MeetingPlatform::GoogleMeet,
-            ),
-            (
-                "https://teams.microsoft.com/v2/",
-                MeetingPlatform::MicrosoftTeams,
-            ),
-            (
-                "https://teams.live.com/meet/123",
-                MeetingPlatform::MicrosoftTeams,
-            ),
-            ("https://app.zoom.us/wc/123", MeetingPlatform::Zoom),
-            (
-                "https://fastrepl.webex.com/meet/test",
-                MeetingPlatform::Webex,
-            ),
-            (
-                "https://app.slack.com/client/workspace/channel",
-                MeetingPlatform::Slack,
-            ),
-        ] {
-            assert_eq!(browser_platform_from_url(Some(url)), Some(platform));
-        }
-
-        for url in [
-            "http://meet.google.com/abc-defg-hij",
-            "https://meet.google.com.evil.example/abc-defg-hij",
-            "https://teams.microsoft.com.evil.example/v2/",
-            "https://zoom.us.evil.example/wc/123",
-            "https://webex.com.evil.example/meet/test",
-            "https://slack.com.evil.example/client/workspace/channel",
-            "javascript:alert(1)",
-        ] {
-            assert_eq!(browser_platform_from_url(Some(url)), None, "accepted {url}");
-        }
-    }
-
-    #[test]
-    fn test_meet_chat_scope_accepts_chromium_webkit_and_gecko_role_variants() {
-        for (container_role, composer_role) in [
-            ("AXGroup", "AXTextArea"),
-            ("AXScrollArea", "AXTextField"),
-            ("AXList", "AXTextArea"),
-        ] {
-            let mut composer = fixture_composer(3, "Send a message", &[1, 0]);
-            composer.role = Some(composer_role.to_string());
-            let nodes = vec![
-                fixture_node(0, "AXWebArea", "Team sync - Google Meet", &[]),
-                fixture_node(1, "AXButton", "Leave call", &[0]),
-                fixture_node(2, container_role, "In-call messages", &[1]),
-                composer,
-            ];
-
-            assert_eq!(
-                validated_chat_scope(&MeetingPlatform::GoogleMeet, &nodes),
-                Some((vec![1], vec![1, 0]))
-            );
-        }
-    }
-
-    #[test]
-    fn test_browser_meeting_window_scope_must_be_unique() {
-        assert_eq!(unique_scope_for_count(0), UniqueMatch::Missing);
-        assert_eq!(unique_scope_for_count(1), UniqueMatch::One(0));
-        assert_eq!(unique_scope_for_count(2), UniqueMatch::Ambiguous);
-        assert_eq!(unique_scope_for_search(1, true), UniqueMatch::One(0));
-        assert_eq!(unique_scope_for_search(1, false), UniqueMatch::Ambiguous);
-    }
-
-    #[test]
-    fn test_webex_native_bundle_classifies_native() {
-        assert_eq!(
-            classify_bundle("Cisco-Systems.Spark"),
-            MeetingPlatform::Webex
-        );
-        assert_eq!(
-            classify_surface("Cisco-Systems.Spark", &MeetingPlatform::Webex),
-            MeetingSurface::Native
-        );
-    }
-
-    #[test]
-    fn test_webex_browser_title_classifies_web() {
-        let web_area = node(21, "AXWebArea", "Cisco Webex Meetings", None);
-        assert_eq!(
-            classify_browser_context(
-                Some("https://fastrepl.webex.com/meet/team"),
-                Some("Cisco Webex Meetings - Brave Browser"),
-                Some(&web_area),
-                &[],
-            ),
-            MeetingPlatform::Webex
-        );
-        assert_eq!(
-            classify_surface("com.brave.Browser", &MeetingPlatform::Webex),
-            MeetingSurface::Web
-        );
-    }
-
-    #[test]
-    fn test_only_provider_like_browser_windows_poison_incomplete_capture() {
-        assert!(!browser_window_has_provider_signal(
-            Some("https://mail.google.com/mail/u/0/#inbox"),
-            Some("Inbox - Gmail"),
-        ));
-        assert!(browser_window_has_provider_signal(
-            Some("https://meet.google.com/abc-defg-hij"),
-            Some("Weekly planning - Google Meet"),
-        ));
-        assert!(browser_window_has_provider_signal(
-            None,
-            Some("Team sync | Microsoft Teams"),
-        ));
-    }
-
-    #[test]
-    fn test_validated_browser_bundles_are_web_surfaces() {
-        for bundle_id in [
-            "com.google.Chrome",
-            "com.microsoft.edgemac",
-            "org.mozilla.firefox",
-            "com.apple.Safari",
-            "com.brave.Browser",
-            "com.vivaldi.Vivaldi",
-            "com.operasoftware.Opera",
-            "company.thebrowser.Browser",
-            "ai.perplexity.comet",
-            "at.studio.AsideBrowser",
-            "company.thebrowser.dia",
-            "com.sigmaos.sigmaos.macos",
-            "net.imput.helium",
-            "com.nousresearch.hermes",
-        ] {
-            assert!(
-                is_browser_bundle(bundle_id),
-                "expected {bundle_id} to be treated as a browser"
-            );
-            assert_eq!(
-                classify_surface(bundle_id, &MeetingPlatform::Zoom),
-                MeetingSurface::Web
-            );
-        }
-    }
-}
+mod tests;
